requirements.txt
anthropic>=0.40.0
streamlit>=1.35.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
numpy>=1.26.0
pandas>=2.2.0
pyspark>=3.5.0
delta-spark>=3.2.0
